refactor: make labeled assertions the required ones
